Create Stone paper scissors
In this code, we use the tkinter module to create a window and various user interface elements, such as buttons and labels. We define a function called choose() that is called whenever the user clicks one of the buttons. This function randomly selects the computer's choice, displays the choices made by the user and the computer, and determines the winner of the game. Finally, we create and pack all the user interface elements into the window, and start the GUI event loop using the mainloop() method

Note that the lambda keyword is used to create anonymous functions that call the choose() function with a specific argument corresponding to the user's choice. This allows us to associate each button with a specific choice without having to create separate functions for each button.